move Version into a separate module
